Explicitly raise TypeError with non-callables
The docs say passing non-callables to 'to' will raise TypeError, but
sometimes a bunch of calls have already happened by the time a
non-callable is found.

This was causing tests to fail under Python 3.7, where some of the
actually callable functions passed were raising other exceptions
before control got the trying to call the non-callables.